refactor: ModalCloseButton 컴포넌트를 Modal.tsx에서 분리
